requirements.txt
transformers
datasets
peft
accelerate
bitsandbytes
torch
torchaudio
soundfile
pyyaml
pytest
gtts
imageio-ffmpeg
fastapi
uvicorn
python-multipart
requests
pydantic
evaluate
tf-keras
scipy
sentencepiece
numpy
pydub
